+++ assets/javascripts/kitten/index.js
@@ -289,6 +289,7 @@ export { RibbonIllustration } from './components/graphics/illustrations/ribbon-i
 export { RainbowIllustration } from './components/graphics/illustrations/rainbow-illustration'
 export { PencilIllustration } from './components/graphics/illustrations/pencil-illustration'
 export { HandAndFlowerIllustration } from './components/graphics/illustrations/hand-and-flower-illustration'
+export { PatchNewIllustration } from './components/graphics/illustrations/patch-new-illustration'
 
 // Graphics/Logos
 export { KissKissBankBankLogo } from './components/graphics/logos/kisskissbankbanklogo'